+++ CHANGELOG.md
@@ -694,6 +694,11 @@
 ### 211210a (4.3.0-post4)
 * Fixed /help message adding a trailing period to the URL
 
+### 220304a (4.3.0-post5)
+* Updated copyright notice for year to 2022
+* IPIDs are no longer generated and stored for clients that do not properly follow the AO protocol when logging in
+* Fixed issue that allowed some handshake commands to be accepted out of order
+
 ### (4.3.1)
 * Added `visible_areas` support to `areas.yaml`. If included in an area, it will set the areas visible to a player in the area. By default it is the same as the reachable areas
 * Added /peek, which allows players to gather information about an area nearby, potentially warning users in the target area of the peek
@@ -702,3 +707,4 @@
   - /paranoia_info
   - /zone_paranoia
   - /zone_paranoia_info
+* Added /pos_force (alias /forcepos), which changes the IC position to a given position of a target player if given, or all players in the current area otherwise